pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chihuahua-tools"
version = "0.1.0"
description = "Chihuahua toolkit (vibe, classification, emoji, horoscope)"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "joblib>=1.3",
  "scikit-learn>=1.3",
]

[project.optional-dependencies]
torch = ["torch", "torchvision", "pillow"]
cv = ["opencv-python", "mediapipe"]
horoscope = ["openai", "python-dotenv"]
dev = ["pytest>=7"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
chihuahua_tools = [
  "assets/models/*",
  "assets/images/*"
]
